# docs/swagger.yaml
basePath: /
definitions:
...
  serializer.UpdatePatientInCalculationRequest:
    properties:
      bread_units:
        type: number
      current_glucose:
        type: number
    required:
    - bread_units
    - current_glucose
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@insulin.ru
    name: API Support
  description: API для расчета болюсного инсулина
  license:
    name: AS IS (NO WARRANTY)
  title: Insulin Calculation API
  version: "1.0"
paths:
...
  /api/insulin-calculations/{insulin_calculation_id}/patients/{patient_id}:
    delete:
      consumes:
      - application/json
      description: Удаляет пациента из расчета инсулина
      parameters:
      - description: ID расчета инсулина
        in: path
        name: insulin_calculation_id
        required: true
        type: integer
      - description: ID пациента
        in: path
        name: patient_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Пациент успешно удален из расчета
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверные ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Расчет или пациент не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить пациента из расчета инсулина
      tags:
      - insulin-calculations
    put:
      consumes:
      - application/json
      description: Обновляет данные пациента в расчете инсулина и пересчитывает инсулин
      parameters:
      - description: ID расчета инсулина
        in: path
        name: insulin_calculation_id
        required: true
        type: integer
      - description: ID пациента
        in: path
        name: patient_id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/serializer.UpdatePatientInCalculationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Данные пациента в расчете успешно обновлены
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Расчет или пациент не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Обновить данные пациента в расчете инсулина
      tags:
      - insulin-calculations
...
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"